Add packet lib and static node codes

+++ myoslib/inc/hal_ultrasonic.h
@@ -2,6 +2,6 @@
 #define HAL_ULTRASONIC_H
 
 void hal_ultrasonic_init();
-double hal_ultrasonic_read();
+uint16_t hal_ultrasonic_read();
 
 #endif
